Client

Description

This module contains functions to parse and build Messages.

Documentation

data Msg

Messages that can be received from the client

Constructors

Plain String

Nick String
Nickname message with the new nick

User String String
User message with username and full name

PrivMsg String String
Private message with receiver and content

Join String
Join message with the channel

JoinMulti [String]
Join message with multiple channels

Part String String
Part message with the channel name and message

Ping String
Ping message

Quit String
Quit with message

IsOn [String]
Is on message with the nick names

parseMsg

Arguments

:: String
Message from the client

-> Msg
Parsed message

This function parses the messages from the client.

buildMsg

Arguments

:: User
The sender of the message

-> Msg
The message

-> String
Message to send to the client

This function builds the messages for the client.

parseCommand

Arguments

:: String
Type of the message

-> String
Remaining message

-> Msg
Parsed message

This function parses the additional information from the message.