Server

Description

This module contains functions for the handling of connections between client and server.

Documentation

data Event

An event that should be handled by the main server

Constructors

ClientMsg User Msg
A parsed message from a client

NewUser User Handle
A new user can bea reached via the given handle

type Users = Map String Handle

This type maps from the nicknames to the handles to the client.

type Channels = Map String [String]

This type maps channel names to the nicknames of the users in the channel.

startServer

Arguments

:: PortNumber
The port that should accept new connections

-> IO ()

This function sets up the listening socket, starts the main server, and starts a loop for accepting connections.

acceptConnections

Arguments

:: Socket
The listening socket

-> Chan Event
The channel to communicate with the main server

-> IO ()

This function accepts new connections from clients and starts the communication with them in a new thread.

mainServer

Arguments

:: Chan Event
The channels to receive events from the clients

-> Users
The currently connected users with their handles

-> Channels
The currently available channels with their users

-> IO ()

The main server waits for events that the clients generate. It also keeps track of the connected users and the channels. It modifies the states and sends replys to the clients.

handleEvent

Arguments

:: Users
The current users

-> Channels
The current channel information

-> Event
The event received from the client

-> IO (Users, Channels)
It returns the new state of users and channels

This function contains the main logic of how to react to the different events. It processes the event, alters the users and the channel information where neccessary and sends responses to the clients.

sendToUser

Arguments

:: Users
The map of users to look up the correct handle

-> String
The nickname of the user to send to

-> String
The message to send

-> IO ()

This function sends a message to a specified user.

sendToAllUsers

Arguments

:: Users
The map of users to look up the correct handles

-> [String]
The nicknames of the users to send to

-> String
The message to send

-> IO ()

This function sends the same message to many users.

startConn

Arguments

:: (Socket, SockAddr)
The socket information for the client

-> Chan Event
The channel to communictate with the main server

-> IO ()

This function sets up the connection to a client and starts the loop that handles the further communication.

loopConn

Arguments

:: Handle
The handle to receive data from the client

-> Chan Event
The channel to communicate with the main server

-> User
The data of the user that is handled by this connection

-> IO ()

This function handles the communication from the client to the server. It reads the messages, parses them, and acts accordingly.

isFinalMsg :: Msg -> Bool

handleMsg :: Handle -> User -> Msg -> (User, Maybe Event)